3 Games in Python
Caterpillar: This game uses turtle module of python. A single stick considered as Caterpillar will be eating a leaf and once it eats, its speed will increase until it gets caught in walls.
EggCatcher:  This game uses tkinter and catches egg that are dropping from the sky by using a basket present on the ground.
MatchMaker: This game uses tkinter and uses the buttons as tiles which are when clicked show their image and then remain opened if they are same images else they will close.